New brands' attributes are fully supported (except API)

+++ app/controllers/Brands.scala
@@ -35,8 +35,6 @@ import play.api.data.Forms._
 import models.UserRole.Role._
 import securesocial.core.SecuredRequest
 import play.api.cache.Cache
-import play.api.data.FormError
-import play.api.data.format.Formatter
 import play.api.libs.json.{ JsValue, Writes, Json }
 import services._
 import fly.play.s3.{ BucketFile, S3Exception }
@@ -73,21 +71,27 @@ object Brands extends Controller with Security {
   def brandsForm(implicit request: SecuredRequest[_]) = Form(mapping(
     "id" -> ignored(Option.empty[Long]),
     "code" -> nonEmptyText.verifying(pattern("[A-Z0-9]*".r, "constraint.brand.code", "constraint.brand.code.error"), maxLength(5)),
+    "uniqueName" -> nonEmptyText.verifying(pattern("[A-Za-z0-9._]*".r, "constraint.brand.code", "constraint.brand.uniqueName.error"), maxLength(25)),
     "name" -> nonEmptyText,
     "coordinatorId" -> nonEmptyText.transform(_.toLong, (l: Long) ⇒ l.toString),
     "description" -> optional(text),
     "picture" -> optional(text),
     "generateCert" -> boolean,
+    "tagLine" -> optional(text),
+    "webSite" -> optional(webUrl),
+    "blog" -> optional(webUrl),
     "profile" -> socialProfileMapping,
     "created" -> ignored(DateTime.now()),
     "createdBy" -> ignored(request.user.fullName),
     "updated" -> ignored(DateTime.now()),
     "updatedBy" -> ignored(request.user.fullName))({
-      (id, code, name, coordinatorId, description, picture, generateCert, profile, created, createdBy, updated, updatedBy) ⇒
-        Brand(id, code, name, coordinatorId, description, picture, generateCert, profile, created, createdBy, updated, updatedBy)
+      (id, code, uniqueName, name, coordinatorId, description, picture, generateCert, tagLine, webSite, blog,
+      profile, created, createdBy, updated, updatedBy) ⇒
+        Brand(id, code, uniqueName, name, coordinatorId, description, picture, generateCert, tagLine, webSite, blog,
+          profile, created, createdBy, updated, updatedBy)
     })({ (b: Brand) ⇒
-      Some((b.id, b.code, b.name, b.coordinatorId, b.description, b.picture, b.generateCert, b.socialProfile, b.created, b.createdBy,
-        b.updated, b.updatedBy))
+      Some((b.id, b.code, b.uniqueName, b.name, b.coordinatorId, b.description, b.picture, b.generateCert, b.tagLine,
+        b.webSite, b.blog, b.socialProfile, b.created, b.createdBy, b.updated, b.updatedBy))
     }))
 
   /** Shows all brands **/
@@ -114,6 +118,9 @@ object Brands extends Controller with Security {
           if (Brand.exists(brand.code))
             Future.successful(BadRequest(views.html.brand.form(request.user, None,
               form.withError("code", "constraint.brand.code.exists", brand.code))))
+          else if (Brand.nameExists(brand.uniqueName))
+            Future.successful(BadRequest(views.html.brand.form(request.user, None,
+              form.withError("uniqueName", "constraint.brand.code.exists", brand.uniqueName))))
           else {
             request.body.asMultipartFormData.get.file("picture").map { picture ⇒
               val filename = Brand.generateImageName(picture.filename)
@@ -187,7 +194,12 @@ object Brands extends Controller with Security {
       }.getOrElse(NotFound)
   }
 
-  /** Edit form submits to this action **/
+  /**
+   * Edit form submits to this action
+   *
+   * @param code Unique text identifier of a brand
+   * @return
+   */
   def update(code: String) = AsyncSecuredRestrictedAction(Editor) { implicit request ⇒
     implicit handler ⇒
 
@@ -196,28 +208,36 @@ object Brands extends Controller with Security {
         form.fold(
           formWithErrors ⇒ Future.successful(BadRequest(views.html.brand.form(request.user, Some(code), form))),
           brand ⇒ {
-            request.body.asMultipartFormData.get.file("picture").map { picture ⇒
-              val filename = Brand.generateImageName(picture.filename)
-              val source = Source.fromFile(picture.ref.file.getPath, encoding)
-              val byteArray = source.toArray.map(_.toByte)
-              source.close()
-              S3Bucket.add(BucketFile(filename, contentType, byteArray)).map { unit ⇒
-                brand.copy(id = existingBrandView.brand.id).copy(picture = Some(filename)).update
-                Cache.remove(Brand.cacheId(code))
-                existingBrandView.brand.picture.map { oldPicture ⇒
-                  S3Bucket.remove(oldPicture)
+            if (Brand.exists(brand.code, existingBrandView.brand.id))
+              Future.successful(BadRequest(views.html.brand.form(request.user, Some(code),
+                form.withError("code", "constraint.brand.code.exists", brand.code))))
+            else if (Brand.nameExists(brand.uniqueName, existingBrandView.brand.id))
+              Future.successful(BadRequest(views.html.brand.form(request.user, Some(code),
+                form.withError("uniqueName", "constraint.brand.code.exists", brand.uniqueName))))
+            else {
+              request.body.asMultipartFormData.get.file("picture").map { picture ⇒
+                val filename = Brand.generateImageName(picture.filename)
+                val source = Source.fromFile(picture.ref.file.getPath, encoding)
+                val byteArray = source.toArray.map(_.toByte)
+                source.close()
+                S3Bucket.add(BucketFile(filename, contentType, byteArray)).map { unit ⇒
+                  brand.copy(id = existingBrandView.brand.id).copy(picture = Some(filename)).update
+                  Cache.remove(Brand.cacheId(code))
+                  existingBrandView.brand.picture.map { oldPicture ⇒
+                    S3Bucket.remove(oldPicture)
+                  }
+                  val activity = Activity.insert(request.user.fullName, Activity.Predicate.Updated, brand.name)
+                  Redirect(routes.Brands.details(code)).flashing("success" -> activity.toString)
+                }.recover {
+                  case S3Exception(status, code, message, originalXml) ⇒
+                    BadRequest(views.html.brand.form(request.user, Some(brand.code),
+                      form.withError("picture", "Image cannot be temporary saved. Please, try again later.")))
                 }
+              }.getOrElse {
+                brand.copy(id = existingBrandView.brand.id).copy(picture = existingBrandView.brand.picture).update
                 val activity = Activity.insert(request.user.fullName, Activity.Predicate.Updated, brand.name)
-                Redirect(routes.Brands.details(code)).flashing("success" -> activity.toString)
-              }.recover {
-                case S3Exception(status, code, message, originalXml) ⇒
-                  BadRequest(views.html.brand.form(request.user, Some(brand.code),
-                    form.withError("picture", "Image cannot be temporary saved. Please, try again later.")))
+                Future.successful(Redirect(routes.Brands.details(code)).flashing("success" -> activity.toString))
               }
-            }.getOrElse {
-              brand.copy(id = existingBrandView.brand.id).copy(picture = existingBrandView.brand.picture).update
-              val activity = Activity.insert(request.user.fullName, Activity.Predicate.Updated, brand.name)
-              Future.successful(Redirect(routes.Brands.details(code)).flashing("success" -> activity.toString))
             }
           })
       }.getOrElse(Future.successful(NotFound))
